+++ src/main/java/dev/provost/BlackjackRunner.java
@@ -1,6 +1,7 @@
 package dev.provost;
 
 import dev.provost.model.Blackjack;
+import dev.provost.model.Card;
 import dev.provost.model.Deck;
 import dev.provost.service.BlackjackService;
 
@@ -10,13 +11,21 @@ public class BlackjackRunner {
 		System.out.println("Hello, welcome to Blackjack.");
 
 		BlackjackService bs = new BlackjackService();
+
+		for (Card c : Deck.getDeckOfCards()) {
+			System.out.println(c);
+		}
+
 		bs.dealCards();
+		bs.clearCards();
 
 		System.out.println(Blackjack.getPlayerHand());
 		System.out.println(Blackjack.getPlayerScore());
 
 		System.out.println(Blackjack.getDealerHand());
 		System.out.println(Blackjack.getDealerScore());
+
+		System.out.println(Blackjack.getDiscardPile());
 
 		System.out.println("Remaining cards: " + Deck.getDeckOfCards().size());
 	}
