+++ src/main/java/dev/provost/model/Blackjack.java
@@ -3,8 +3,6 @@ package dev.provost.model;
 import java.util.ArrayList;
 import java.util.List;
 
-import dev.provost.model.Card.Suit;
-
 public class Blackjack {
 
 	private static List<Card> dealerHand = new ArrayList<>();
@@ -12,7 +10,6 @@ public class Blackjack {
 	private static List<Card> discardPile = new ArrayList<>();
 	private static int dealerScore = 0;
 	private static int playerScore = 0;
-	private static Deck deckOfCards = new Deck();
 
 	public static List<Card> getDealerHand() {
 		return dealerHand;
@@ -62,19 +59,4 @@ public class Blackjack {
 		}
 	}
 
-//	public static Deck getDeck() {
-//		return deckOfCards;
-//	}
-
-	public static void setDeck(Deck deck) {
-		Blackjack.deckOfCards = deck;
-	}
-
-	public static void replaceAceInHand(int index, Suit suit, int score) {
-		if (playerHand.get(index).getValue().equals("Ace")) {
-			playerHand.set(index, new Card("Ace", suit, score));
-		} else {
-			System.out.println("No Ace in that position.");
-		}
-	}
 }
