+++ src/main/java/dev/provost/service/BlackjackService.java
@@ -1,5 +1,7 @@
 package dev.provost.service;
 
+import java.util.List;
+
 import dev.provost.model.Blackjack;
 import dev.provost.model.Card;
 import dev.provost.repository.BlackjackDAO;
@@ -26,7 +28,23 @@ public class BlackjackService {
 		for (Card c : Blackjack.getDealerHand()) {
 			bd.addToDiscardPile(c);
 		}
-		Blackjack.getPlayerHand().removeAll(null);
+		Blackjack.getPlayerHand().removeAll(Blackjack.getPlayerHand());
+		Blackjack.getDealerHand().removeAll(Blackjack.getDealerHand());
+	}
+
+	public void hitMove(List<Card> hand) {
+		hand.add(bd.drawCard());
+	}
+
+	public void changeAceScore(int index) {
+		Card c = Blackjack.getPlayerHand().get(index);
+		if (c.getScore() == 1) {
+			bd.replaceAceInHand(index, c.getSuit(), 11);
+		} else if (c.getScore() == 11) {
+			bd.replaceAceInHand(index, c.getSuit(), 1);
+		} else {
+			System.out.println("No Aces to change.");
+		}
 	}
 
 }
